// src/store/modules/bandSync/action-types.ts
export enum BandSyncActionTypes {
    INIT = "BAND_SYNC_INIT",
    CREATE_SESSION = "BAND_SYNC_CREATE_SESSION",
    JOIN_SESSION = "BAND_SYNC_JOIN_SESSION",
    CANCEL_START = "BAND_SYNC_CANCEL_START",
    LEAVE_SESSION = "BAND_SYNC_LEAVE_SESSION",
    LEADER_PUBLISH = "BAND_SYNC_LEADER_PUBLISH",
    SET_MEMBER_SYNC_STATUS = "BAND_SYNC_SET_MEMBER_SYNC_STATUS",
    MEMBER_SYNCED = "BAND_SYNC_MEMBER_SYNCED",
    SESSION_SNAPSHOT = "BAND_SYNC_SESSION_SNAPSHOT",
    SESSION_ENDED = "BAND_SYNC_SESSION_ENDED",
}